requirements.txt
# See-Think-Act AI Agent Requirements
# For Windows 11

# Core dependencies
mss>=9.0.1                    # Fast cross-platform screen capture
pyautogui>=0.9.54            # GUI automation (mouse & keyboard)
Pillow>=10.0.0               # Image processing
ollama>=0.1.0                # Ollama Python client

# Qwen Agent dependencies
qwen-agent>=0.0.5            # Qwen agent framework for function calling

# Additional utilities
python-dotenv>=1.0.0         # Environment variable management

# Jupyter support (optional)
ipykernel>=6.25.0
jupyter>=1.0.0
ipython>=8.12.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
